Move DataQueryResult -> ViewerContext
